Update qmath.cpp
Added a differentiation function, supply the function head and the point of differentiation and der will return the derivative of the interpolated 4th degree polynomial.
Added a numerical integration function. Supply the function head, an arbitrary variable, and a lower and upper bound to obtain your integral.
Added the user controlled QMATHDELTA for the calculus functions, this determines sensitivity for integration and differentiation. I'm looking into how to do nested integrals and derivatives.